hands_on_microservices/backend/requirements.txt
fastapi==0.95.2
uvicorn==0.22.0
psycopg2-binary==2.9.6
sqlalchemy==2.0.15
alembic==1.11.1
pydantic==1.10.7
python-jose==3.3.0 
passlib==1.7.4
Pillow==9.5.0
numpy==1.24.2
scikit-learn==1.2.2
python-multipart
transformers
torch